Updated README.md for setup

+++ purge_all.sh
@@ -1,9 +1,10 @@
 #! /bin/bash
 
-rm -r -f ethereum accounts.txt enode.txt;
+[[ -d ethereum ]] && rm -r -f ethereum;
+[[ -f accounts.txt ]] && [[ -f enode.txt ]] && rm accounts.txt enode.txt;
 
 for i in 1 2 3 4 5
 do
-    rm "start_node_$i.sh";
+    [[ -f "start_node_$i.sh" ]] && rm "start_node_$i.sh";
 
 done
